requirements.txt
antlr4-python3-runtime==4.9.3
cycler==0.11.0
dataclasses==0.8
hydra-core==1.2.0
importlib-resources==5.4.0
kiwisolver==1.3.1
matplotlib==3.3.4
numpy==1.19.5
omegaconf==2.2.3
packaging==21.3
Pillow==8.4.0
pyparsing==3.0.9
python-dateutil==2.8.2
PyYAML==6.0
six==1.16.0
torch==1.10.2
typing_extensions==4.1.1
zipp==3.6.0
